requirements.txt
absl-py==2.0.0
addict==2.4.0
attrs==23.1.0
black==23.11.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.1.1
cycler==0.12.1
descartes==1.1.0
efficientnet-pytorch==0.7.1
exceptiongroup==1.2.0
filelock==3.13.1
fiona==1.9.5
fire==0.5.0
flake8==6.1.0
fonttools==4.45.1
fsspec==2023.12.1
geopandas==0.13.2
google-auth==2.23.4
google-auth-oauthlib==1.0.0
grpcio==1.59.3
huggingface-hub==0.19.4
idna==3.6
imageio==2.33.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==2.0.0
ipython==8.12.2
jedi==0.19.1
joblib==1.3.2
kiwisolver==1.4.5
llvmlite==0.36.0
lyft-dataset-sdk==0.0.8
markdown==3.5.1
markupsafe==2.1.3
matplotlib==3.5.3
mccabe==0.7.0
mmcls==0.25.0
mmcv-full==1.6.0
mmdet==2.28.2
mmsegmentation==0.30.0
mypy-extensions==1.0.0
networkx==2.2
numba==0.53.0
numpy==1.23.5
nuscenes-devkit==1.1.11
nvidia-ml-py==12.535.133
nvitop==1.3.1
oauthlib==3.2.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.0.3
pathspec==0.11.2
pillow==10.1.0
platformdirs==4.0.0
plotly==5.18.0
pluggy==1.3.0
plyfile==1.0.2
prettytable==3.9.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycodestyle==2.11.1
pyflakes==3.1.0
pyparsing==3.1.1
pyproj==3.5.0
pyquaternion==0.9.9
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
pywavelets==1.4.1
pyyaml==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-image==0.19.3
scikit-learn==1.3.2
scipy==1.10.1
shapely==1.8.5.post1
some-package==0.1
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboardx==2.6.2.2
termcolor==2.3.0
terminaltables==3.1.10
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.12
tomli==2.0.1
torch-scatter==2.0.9
tqdm==4.66.1
trimesh==2.35.39
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.12
werkzeug==3.0.1
yapf==0.40.2
zipp==3.17.0